refactor(filament): merge traits into concerns, drop dead code and deprecated APIs
- move TranslatableResource from app/Filament/Traits to app/Filament/Concerns
- extract the copy-pasted truncated-tooltip closure into TruncatedTooltip
- rename getActions() to getHeaderActions() and ->actions()/->bulkActions()
  to ->recordActions()/->toolbarActions()
- remove v2-era Filament\Pages\Actions imports
- delete unregistered CatalogEditors CRUD pages, DateRangeFilter,
  EditWebsiteAd, unused Login form schema, getPrimaryKey() stubs,
  BadgeUploadResource table helpers and two unreferenced blade views
