+++ app/Filament/Concerns/TranslatableResource.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Filament\Traits;
+namespace App\Filament\Concerns;
 
 use Str;
