+++ app/Filament/Resources/Atom/Articles/Pages/ListArticles.php
@@ -10,7 +10,7 @@ class ListArticles extends ListRecords
 {
     protected static string $resource = ArticleResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
