+++ app/Filament/Resources/Atom/Tags/Pages/EditTag.php
@@ -10,7 +10,7 @@ class EditTag extends EditRecord
 {
     protected static string $resource = TagResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             DeleteAction::make(),
