+++ app/Filament/Resources/Atom/Tags/Pages/ListTags.php
@@ -10,7 +10,7 @@ class ListTags extends ListRecords
 {
     protected static string $resource = TagResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
