+++ app/Filament/Resources/Atom/Teams/Pages/EditTeam.php
@@ -10,7 +10,7 @@ class EditTeam extends EditRecord
 {
     protected static string $resource = TeamResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             DeleteAction::make(),
