+++ app/Filament/Resources/Atom/Teams/Pages/ListTeams.php
@@ -10,7 +10,7 @@ class ListTeams extends ListRecords
 {
     protected static string $resource = TeamResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
