+++ app/Filament/Resources/Home/HomeCategoryResource/Pages/EditHomeCategory.php
@@ -12,7 +12,7 @@ class EditHomeCategory extends EditRecord
 {
     protected static string $resource = HomeCategoryResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             DeleteAction::make()
