+++ app/Filament/Resources/Home/HomeCategoryResource/Pages/ListHomeCategories.php
@@ -10,7 +10,7 @@ class ListHomeCategories extends ListRecords
 {
     protected static string $resource = HomeCategoryResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
