+++ app/Filament/Resources/Home/HomeItemResource/Pages/EditHomeItem.php
@@ -12,7 +12,7 @@ class EditHomeItem extends EditRecord
 {
     protected static string $resource = HomeItemResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             DeleteAction::make()
