+++ app/Filament/Resources/Home/HomeItemResource/Pages/ListHomeItems.php
@@ -10,7 +10,7 @@ class ListHomeItems extends ListRecords
 {
     protected static string $resource = HomeItemResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
