+++ app/Filament/Resources/Hotel/Achievements/Pages/ListAchievements.php
@@ -3,17 +3,14 @@
 namespace App\Filament\Resources\Hotel\Achievements\Pages;
 
 use App\Filament\Resources\Hotel\Achievements\AchievementResource;
-use Filament\Pages\Actions;
 use Filament\Resources\Pages\ListRecords;
 
 class ListAchievements extends ListRecords
 {
     protected static string $resource = AchievementResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
-        return [
-            // Actions\CreateAction::make(),
-        ];
+        return [];
     }
 }
