+++ app/Filament/Resources/Hotel/BadgeTextEditors/Pages/EditBadgeTextEditor.php
@@ -14,17 +14,10 @@ class EditBadgeTextEditor extends EditRecord
 {
     protected static string $resource = BadgeTextEditorResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [DeleteAction::make()];
     }
-
-    protected function mutateFormDataBeforeSave(array $data): array
-    {
-        return $data;
-    }
-
-    protected function afterSave(): void {}
 
     protected function handleRecordUpdate(Model $record, array $data): Model
     {
