+++ app/Filament/Resources/Hotel/BadgeTextEditors/Pages/ListBadgeTextEditors.php
@@ -16,7 +16,7 @@ class ListBadgeTextEditors extends ListRecords
 {
     protected static string $resource = BadgeTextEditorResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make()
