--- app/Filament/Resources/Hotel/CatalogEditors/Pages/CreateCatalogEditor.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Hotel\CatalogEditors\Pages;
-
-use App\Filament\Resources\Hotel\CatalogEditors\CatalogEditorResource;
-use Filament\Resources\Pages\CreateRecord;
-
-class CreateCatalogEditor extends CreateRecord
-{
-    protected static string $resource = CatalogEditorResource::class;
-}
