+++ app/Filament/Resources/Hotel/CommandLogs/Pages/ManageCommandLogs.php
@@ -9,13 +9,8 @@ class ManageCommandLogs extends ManageRecords
 {
     protected static string $resource = CommandLogResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [];
-    }
-
-    public function getPrimaryKey(): string
-    {
-        return 'timestamp';
     }
 }
