+++ app/Filament/Resources/Hotel/EmulatorSettings/EmulatorSettingResource.php
@@ -2,11 +2,13 @@
 
 namespace App\Filament\Resources\Hotel\EmulatorSettings;
 
+use App\Emulator\Data\Feature;
 use App\Filament\Concerns\RequiresEmulatorDriver;
+use App\Filament\Concerns\RequiresEmulatorFeature;
+use App\Filament\Concerns\TranslatableResource;
 use App\Filament\Resources\Hotel\EmulatorSettings\Pages\CreateEmulatorSetting;
 use App\Filament\Resources\Hotel\EmulatorSettings\Pages\EditEmulatorSetting;
 use App\Filament\Resources\Hotel\EmulatorSettings\Pages\ListEmulatorSettings;
-use App\Filament\Traits\TranslatableResource;
 use App\Models\EmulatorSetting;
 use Filament\Actions\EditAction;
 use Filament\Actions\ViewAction;
