+++ app/Filament/Resources/Hotel/EmulatorSettings/Pages/EditEmulatorSetting.php
@@ -10,7 +10,7 @@ class EditEmulatorSetting extends EditRecord
 {
     protected static string $resource = EmulatorSettingResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             DeleteAction::make(),
