+++ app/Filament/Resources/Hotel/EmulatorSettings/Pages/ListEmulatorSettings.php
@@ -10,7 +10,7 @@ class ListEmulatorSettings extends ListRecords
 {
     protected static string $resource = EmulatorSettingResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
