+++ app/Filament/Resources/Hotel/EmulatorTexts/EmulatorTextResource.php
@@ -2,9 +2,11 @@
 
 namespace App\Filament\Resources\Hotel\EmulatorTexts;
 
+use App\Emulator\Data\Feature;
 use App\Filament\Concerns\RequiresEmulatorDriver;
+use App\Filament\Concerns\RequiresEmulatorFeature;
+use App\Filament\Concerns\TranslatableResource;
 use App\Filament\Resources\Hotel\EmulatorTexts\Pages\ManageEmulatorTexts;
-use App\Filament\Traits\TranslatableResource;
 use App\Models\EmulatorText;
 use Filament\Actions\BulkActionGroup;
 use Filament\Actions\DeleteBulkAction;
