--- app/Filament/Resources/Hotel/WebsiteAds/Pages/EditWebsiteAd.php
@@ -1,19 +1,0 @@
-<?php
-
-namespace App\Filament\Resources\Hotel\WebsiteAds\Pages;
-
-use App\Filament\Resources\Hotel\WebsiteAds\WebsiteAdResource;
-use Filament\Actions\DeleteAction;
-use Filament\Resources\Pages\EditRecord;
-
-class EditWebsiteAd extends EditRecord
-{
-    protected static string $resource = WebsiteAdResource::class;
-
-    protected function getHeaderActions(): array
-    {
-        return [
-            DeleteAction::make(),
-        ];
-    }
-}
