+++ app/Filament/Resources/User/Bans/Pages/ManageBans.php
@@ -9,7 +9,7 @@ class ManageBans extends ManageRecords
 {
     protected static string $resource = BanResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             // ...
