+++ app/Filament/Resources/User/Users/Pages/EditUser.php
@@ -24,7 +24,7 @@ class EditUser extends EditRecord
 {
     protected static string $resource = UserResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             DeleteAction::make(),
