+++ app/Filament/Resources/User/Users/Pages/ListUsers.php
@@ -10,7 +10,7 @@ class ListUsers extends ListRecords
 {
     protected static string $resource = UserResource::class;
 
-    protected function getActions(): array
+    protected function getHeaderActions(): array
     {
         return [
             CreateAction::make(),
