+++ app/Filament/Resources/User/Users/UserResource.php
@@ -6,12 +6,12 @@ use App\Emulator\Contracts\PlayerSettingsRepository;
 use App\Emulator\Contracts\RankRepository;
 use App\Emulator\Data\Feature;
 use App\Emulator\Emulator;
+use App\Filament\Concerns\TranslatableResource;
 use App\Filament\Resources\User\Users\Pages\CreateUser;
 use App\Filament\Resources\User\Users\Pages\EditUser;
 use App\Filament\Resources\User\Users\Pages\ListUsers;
 use App\Filament\Resources\User\Users\Pages\ViewUser;
 use App\Filament\Tables\Columns\UserAvatarColumn;
-use App\Filament\Traits\TranslatableResource;
 use App\Models\Community\Staff\WebsiteTeam;
 use App\Models\User;
 use Filament\Actions\EditAction;
